Implemented precheck to be sure stack parameters exist before running program.  All Unit Tests pass.
