+++ SOS/Plush/Plush.BooleanInstructions.cpp
@@ -20,17 +20,32 @@ namespace Plush
 		//make_instruction((Operator)inallrev<bool>, "BOOLEAN.INALLREV", nullType, boolType);
 		//make_instruction((Operator)out<bool>, "BOOLEAN.OUT", integerBool, nullType);
 
-		make_instruction((Operator)_and, "BOOLEAN", "AND");
-		make_instruction((Operator)_or, "BOOLEAN", "OR");
-		make_instruction((Operator)_not, "BOOLEAN", "NOT");
-		make_instruction((Operator)_nand, "BOOLEAN", "NAND");
-		make_instruction((Operator)_nor, "BOOLEAN", "NOR");
-		make_instruction((Operator)float2bool, "BOOLEAN", "FROMFLOAT");
-		make_instruction((Operator)int2bool, "BOOLEAN", "FROMINTEGER");
-		make_instruction((Operator)in<bool>, "BOOLEAN", "IN");
-		make_instruction((Operator)inall<bool>, "BOOLEAN", "INALL");
-		make_instruction((Operator)inallrev<bool>, "BOOLEAN", "INALLREV");
-		make_instruction((Operator)out<bool>, "BOOLEAN", "OUT");
+		//make_instruction((Operator)_and, "BOOLEAN", "AND");
+		//make_instruction((Operator)_or, "BOOLEAN", "OR");
+		//make_instruction((Operator)_not, "BOOLEAN", "NOT");
+		//make_instruction((Operator)_nand, "BOOLEAN", "NAND");
+		//make_instruction((Operator)_nor, "BOOLEAN", "NOR");
+		//make_instruction((Operator)float2bool, "BOOLEAN", "FROMFLOAT");
+		//make_instruction((Operator)int2bool, "BOOLEAN", "FROMINTEGER");
+		//make_instruction((Operator)in<bool>, "BOOLEAN", "IN");
+		//make_instruction((Operator)inall<bool>, "BOOLEAN", "INALL");
+		//make_instruction((Operator)inallrev<bool>, "BOOLEAN", "INALLREV");
+		//make_instruction((Operator)out<bool>, "BOOLEAN", "OUT");
+
+
+		Type binaryBool = boolType + boolType;
+
+		push_make_instruction((Operator)_and, "BOOLEAN", "AND", binaryBool);
+		push_make_instruction((Operator)_or, "BOOLEAN", "OR", binaryBool);
+		push_make_instruction((Operator)_not, "BOOLEAN", "NOT", boolType);
+		push_make_instruction((Operator)_nand, "BOOLEAN", "NAND", binaryBool);
+		push_make_instruction((Operator)_nor, "BOOLEAN", "NOR", binaryBool);
+		push_make_instruction((Operator)float2bool, "BOOLEAN", "FROMFLOAT", floatType);
+		push_make_instruction((Operator)int2bool, "BOOLEAN", "FROMINTEGER", integerType);
+		push_make_instruction((Operator)in<bool>, "BOOLEAN", "IN", integerType);
+		push_make_instruction((Operator)inall<bool>, "BOOLEAN", "INALL", nullType);
+		push_make_instruction((Operator)inallrev<bool>, "BOOLEAN", "INALLREV", nullType);
+		push_make_instruction((Operator)out<bool>, "BOOLEAN", "OUT", out_in<bool>());
 	}
 
 }
