+++ SOS/Plush/Plush.FloatInstructions.h
@@ -21,7 +21,8 @@ namespace Plush
 	inline unsigned _fmod(Environment & _env)
 	{
 		// Check for valid parameters
-		if ((_env.has_elements<double>(2)) && (_env.get_top<double>() != 0.0))
+		//if ((_env.has_elements<double>(2)) && (_env.get_top<double>() != 0.0))
+		if (_env.get_top<double>() != 0.0)
 		{
 			double first = _env.pop<double>();
 			double second = _env.pop<double>();
@@ -34,7 +35,7 @@ namespace Plush
 
 	inline unsigned _cos(Environment & _env)
 	{
-		if (_env.has_elements<double>(1))
+		//if (_env.has_elements<double>(1))
 			_env.set_top<double>(cos(_env.get_top<double>()));
 
 		return 1;
@@ -42,7 +43,7 @@ namespace Plush
 
 	inline unsigned _sin(Environment & _env)
 	{
-		if (_env.has_elements<double>(1))
+		//if (_env.has_elements<double>(1))
 			_env.set_top<double>(sin(_env.get_top<double>()));
 
 		return 1;
@@ -50,7 +51,7 @@ namespace Plush
 
 	inline unsigned _tan(Environment & _env)
 	{
-		if (_env.has_elements<double>(1))
+		//if (_env.has_elements<double>(1))
 		{
 			double val = _env.get_top<double>();
