+++ SOS/Plush/Plush.StaticInit.cpp
@@ -1,4 +1,6 @@
 #include "Plush.StaticInit.h"
+#include "Environment.h"
+#include "Plush.Instruction.h"
 
 namespace Plush
 {
@@ -15,8 +17,26 @@ namespace Plush
 	void initFloat();
 	void initInt();
 
+	Type execType; //(EXEC_STACK+1,EXEC_STACK,1);
+	Type integerType; //(INTEGER_STACK+1,INTEGER_STACK,1);
+	Type codeType; //(CODE_STACK+1, CODE_STACK,1);
+	Type boolType; //(BOOL_STACK+1, BOOL_STACK,1);
+	Type floatType; //(FLOAT_STACK+1, FLOAT_STACK,1);
+	Type nullType;
+
+	void init_types()
+	{
+		execType = Type(EXEC_STACK + 1, EXEC_STACK, 1);
+		integerType = Type(INTEGER_STACK + 1, INTEGER_STACK, 1);
+		codeType = Type(CODE_STACK + 1, CODE_STACK, 1);
+		boolType = Type(BOOL_STACK + 1, BOOL_STACK, 1);
+		floatType = Type(FLOAT_STACK + 1, FLOAT_STACK, 1);
+	}
+
 	StaticInit::StaticInit()
 	{
+		init_types();
+
 		initGenerics();
 		initBool();
 //		initCode();
@@ -25,12 +45,19 @@ namespace Plush
 		initInt();
 	};
 
-	void StaticInit::register_pushfunc(Operator op, std::string type, std::string name)
+	void StaticInit::push_register_pushfunc(Instruction* pInstruction)
 	{
-		std::string func_name = type + "." + name;
-		Func2CodeMap[func_name] = op;
+		std::string func_name = pInstruction->to_string();
+		Func2CodeMap[func_name] = pInstruction;
 		function_names.push_back(func_name);
-	};
+	}
+
+	//void StaticInit::register_pushfunc(Operator op, std::string type, std::string name)
+	//{
+	//	std::string func_name = type + "." + name;
+	//	Func2CodeMap[func_name] = op;
+	//	function_names.push_back(func_name);
+	//};
 
 	void StaticInit::set_parentheses(std::string type, std::string name, unsigned int block_wants)
 	{
@@ -52,5 +79,10 @@ namespace Plush
 	std::string StaticInit::get_function_name(unsigned int function_index)
 	{
 		return function_names[function_index];
+	}
+
+	void push_make_instruction(Operator op, std::string type, std::string name, Type in)
+	{
+		static_initializer.push_register_pushfunc(new Instruction(op, type + "." + name, in));
 	}
 }
