+++ SOS/Plush/Plush.StaticInit.h
@@ -1,13 +1,22 @@
 #pragma once
-#include <string>
 #include <map>
-#include "Processor.h"
+#include <string>
+#include <vector>
+#include "Plush.Instruction.h"
+//#include "Processor.h"
 
 namespace Plush
 {
+	class Instruction;
+
 	typedef unsigned(*Operator)(Environment &env);
 
-	typedef std::map<std::string, Operator> Func2CodeMapType;
+	typedef std::map<std::string, unsigned int> Func2BlockWantsMapType;
+	extern Func2BlockWantsMapType Func2BlockWantsMap;
+	
+//	typedef std::map<std::string, Operator> Func2CodeMapType;
+
+	typedef std::map<std::string, Instruction*> Func2CodeMapType;
 	extern 	Func2CodeMapType Func2CodeMap;
 
 //	typedef std::vector<std::string> Names;
@@ -27,7 +36,8 @@ namespace Plush
 		StaticInit();
 
 		/* Registers a function and makes it globally available through the 'instructions' Code */
-		void register_pushfunc(Operator op, std::string type, std::string name);
+		void push_register_pushfunc(Instruction* pInstruction);
+		//void register_pushfunc(Operator op, std::string type, std::string name);
 		void set_parentheses(std::string type, std::string name, unsigned int block_wants);
 		void set_parentheses(std::string name, unsigned int block_wants);
 
@@ -37,10 +47,10 @@ namespace Plush
 
 	extern StaticInit static_initializer;
 
-	inline void make_instruction(Operator op, std::string type, std::string name)
-	{
-		static_initializer.register_pushfunc(op, type, name);
-	}
+	//inline void make_instruction(Operator op, std::string type, std::string name)
+	//{
+	//	static_initializer.register_pushfunc(op, type, name);
+	//}
 
 	inline void set_parentheses(std::string type, std::string name, unsigned int block_wants)
 	{
@@ -51,4 +61,6 @@ namespace Plush
 	{
 		static_initializer.set_parentheses(name, block_wants);
 	}
+
+	void push_make_instruction(Operator op, std::string type, std::string name, Type in);
 }
