+++ UnitTest/UnitTest.vcxproj
@@ -97,15 +97,15 @@
       <WarningLevel>Level3</WarningLevel>
       <Optimization>Disabled</Optimization>
       <SDLCheck>true</SDLCheck>
-      <AdditionalIncludeDirectories>C:\Users\egarrity\Documents\Source\Repo\SOS\SOS\Plush;C:\Users\edgar\OneDrive\Projects\Source\Repos\EdGarrity\SOS\SOS\Plush;$(VCInstallDir)UnitTest\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
+      <AdditionalIncludeDirectories>C:\Users\edgar\OneDrive\Projects\Source\Repos\EdGarrity\SOS\SOS\Plush;$(VCInstallDir)UnitTest\include;%(AdditionalIncludeDirectories)</AdditionalIncludeDirectories>
       <PreprocessorDefinitions>_DEBUG;%(PreprocessorDefinitions)</PreprocessorDefinitions>
       <UseFullPaths>true</UseFullPaths>
       <PrecompiledHeaderFile>pch.h</PrecompiledHeaderFile>
     </ClCompile>
     <Link>
       <SubSystem>Windows</SubSystem>
-      <AdditionalLibraryDirectories>C:\Users\egarrity\source\repos\EdGarrity\SOS\SOS\x64\Debug;C:\Users\egarrity\Documents\Source\Repo\SOS\SOS\x64\Debug;C:\Users\egarrity\OneDrive\Projects\Source\Repos\EdGarrity\SOS\SOS\x64\Debug;C:\Users\edgar\OneDrive\Projects\Source\Repos\EdGarrity\SOS\SOS\x64\Debug;$(VCInstallDir)UnitTest\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
-      <AdditionalDependencies>Atom.obj;Environment.obj;Genome;Plush.BooleanInstructions.obj;Plush.ExecInstruction.obj;Plush.FloatInstructions.obj;Plush.IntegerInstructions.obj;Plush.Instruction.obj;Plush.StaticInit.obj;Processor.obj;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalLibraryDirectories>C:\Users\edgar\OneDrive\Projects\Source\Repos\EdGarrity\SOS\SOS\x64\Debug;$(VCInstallDir)UnitTest\lib;%(AdditionalLibraryDirectories)</AdditionalLibraryDirectories>
+      <AdditionalDependencies>AsyncBreed.obj;Atom.obj;AtomicFunctions.obj;Breed.obj;Conversion.obj;Environment.obj;ErrorFunction.LearnFromExample.obj;FixedSizeStack.obj;GeneticOperators.obj;Genome.obj;Globals.obj;Individual.obj;main.obj;MyException.obj;pch.obj;Plush.BooleanInstructions.obj;Plush.ExecInstruction.obj;Plush.FloatInstructions.obj;Plush.Instruction.obj;Plush.IntegerInstructions.obj;Plush.StaticInit.obj;Processor.obj;Random.obj;Random.Utilities.obj;Run.LearnFromExamples.obj;Selection.obj;SimulatedAnnealing.obj;SQLCommand.obj;SQLConnection.obj;String.obj;SystemInfo.obj;Type.obj;Utilities.obj;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
   </ItemDefinitionGroup>
   <ItemDefinitionGroup Condition="'$(Configuration)|$(Platform)'=='Debug|Win32'">
@@ -171,7 +171,8 @@
     <ClCompile Include="BooleanInstructions.cpp" />
     <ClCompile Include="CodeInstructions.cpp" />
     <ClCompile Include="ExecInstructions.cpp" />
-    <ClCompile Include="StackInstructions.cpp" />
+    <ClCompile Include="FloatInstructions.cpp" />
+    <ClCompile Include="IntegerInstructions.cpp" />
     <ClCompile Include="UnitTest.cpp" />
     <ClCompile Include="Utilities\HelperFunctions.cpp" />
   </ItemGroup>
